update viz; fix types issue

+++ aerosandbox/dynamics/point_mass/common_point_mass.py
@@ -477,7 +477,7 @@ class _DynamicsPointMassBaseClass(AeroSandboxObject, ABC):
                         (np.diff(x_e) ** 2 + np.diff(y_e) ** 2 + np.diff(z_e) ** 2) ** 0.5
                     )
                     vehicle_length = np.diff(vehicle_bounds[0, :])
-                    scale_vehicle_model = 0.5 * path_length / vehicle_length / n_vehicles_to_draw
+                    scale_vehicle_model = float(0.5 * path_length / vehicle_length / n_vehicles_to_draw)
 
             ### Initialize the plotter
             if plotter is None:
